requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pillow
PyJWT
firebase-admin
onnxruntime
faiss-cpu
motor